full/ml-model-custom-vision/src/requirements.txt
 python-dotenv
 azure-cognitiveservices-vision-customvision
 